Extraer feedback sonoro de mark.js a un módulo propio
Segundo paso de la descomposición de mark.js: getAudioCtx(),
playTone() y playBeep() se mueven a mark/audio-feedback.js, junto con
el estado userHasInteracted que solo existía para gatear el audio (Web
Audio API bloquea sonido hasta el primer gesto del usuario). El módulo
expone markUserInteracted() y playBeep() — mark.js llama a la primera
en los 4 puntos donde antes seteaba la variable local. Refactor puro,
sin cambios de comportamiento.
